move getParams
