+++ src/sentry/static/sentry/app/components/organizations/globalSelectionHeader/getParams.tsx
